hybrid-rag-system/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
langchain>=0.3.7
langchain-openai>=0.2.9
langchain-community>=0.3.7
langchain-text-splitters>=0.3.2
chromadb>=1.0.0
rank-bm25>=0.2.2
cohere>=5.11.4
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.1
nltk>=3.9.1
httpx>=0.27.2
pypdf>=4.3.1
python-docx>=1.1.2
beautifulsoup4>=4.12.3
lxml>=5.3.0
streamlit>=1.40.0
